# go_app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Agendly",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/review": {
            "post": {
                "description": "Create a review for a shop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "AddReview",
                "parameters": [
                    {
                        "description": "shop id_shop",
                        "name": "shop",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "user id_user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "review rating",
                        "name": "rating",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "review comment",
                        "name": "comment",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/review/{id}": {
            "get": {
                "description": "Retrieve all reviews for a shop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "GetReviewByShop",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "shop id_shop",
                        "name": "shop",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Delete a review for a shop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "DeleteReview",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "review id_review",
                        "name": "review",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "description": "Update a review for a shop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "UpdateReview",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "review id_review",
                        "name": "review",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "review rating",
                        "name": "rating",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "review comment",
                        "name": "comment",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}